Allowing a custom UNIT_CMD and INTEGRATION_CMD
Instead of assuming you're using Karma and Protractor it lets you specify your command. This is useful even if you are using Karma and Protractor because you might have a different command than the default.
