+++ index.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <title>Example Test</title>
+  </head>
+  <body>
+    <h1>Example Test</h1>
+    <p>
+      This page is loaded with Protractor as an example test.
+    </p>
+  </body>
+</html>
