+++ tests/integration/example.js
@@ -1,0 +1,9 @@
+describe('Example Test', function() {
+  beforeEach(function() {
+    browser.ignoreSynchronization = true;
+    browser.get('');
+  });
+  it('has the right title', function() {
+    expect(browser.getTitle()).toEqual('Example Test');
+  });
+});
