+++ tests/unit/example.js
@@ -1,0 +1,5 @@
+describe('example test', function() {
+  it('works', function() {
+    expect(true).toBe(true);
+  });
+});
